New excel table, divided result handling functiion
I changed the excel_processing so it fits the new Excel table that only includes all relevant data for the re-analysis.

I also divided the "extract_coordinates_from_pdfs_in_folder" method into two seperate functions, one that handles the coordinates and its context lines only and one that is only there for the result handling so its way easier to add new results to it. For the decision logic (differentiate between coordinates and no coordinates as main differrence) we still have to call "extract_coordinates_from_pdfs_in_folder" in the main