Docker, mayor module logic changes, translations
Employed Docker, changed the main module logic to work with the 'process_extraction_results' function so the 'extract_spatial_information_from_pdfs' function is only for extracting spatial information. Also started translation comments and Docstrings to keep it uniform with the other mayor python script 'Creating_plots.py'

Also deleted the forest ecosystem search and extraction, since it is not needed for the bachelor thesis (anymore).